requirements.txt
# Data ingestion and manipulation
pandas
datasets
huggingface-hub
tqdm

# Vector database and storage
qdrant-client==1.9.0

# Embeddings and NLP
sentence-transformers>=3.0.0

# LLM API and enviroment
openai==1.50.0
python-dotenv

# Development and notebook
ipykernel
ipywidgets






